Implement new message format in addition to the old one, which is now deprecated but will be supported in this version.
This commit implements new features for generating messages of the newly defined format, but will still support the old TriangleMesh message for an easier transition between the two message formats. Users of this node should quickly adapt to the new format, since it is not intended to keep supporting the old TriangleMesh format for a longer period of time.